feat: fix

+++ src/main/java/leetcode/problems/LeetCode437_pre.java
@@ -1,12 +1,13 @@
 package leetcode.problems;
 
-import common.TreeNode;
+import common.*;
 
-import java.util.HashMap;
+import java.util.*;
 
 /**
  * @author wujingxinit@outlook.com
  * @date 2023/8/14 18:10
+ * {@link LeetCode560}
  */
 public class LeetCode437_pre {
 
@@ -20,20 +21,19 @@ public class LeetCode437_pre {
         this.targetSum = targetSum;
         HashMap<Long, Integer> map = new HashMap<>();
         map.put(0L, 1);
-        return (int) dfs(map, root, 0);
+        return dfs(map, root, 0);
     }
 
-    public long dfs(HashMap<Long, Integer> map, TreeNode root, long sum) {
+    public int dfs(HashMap<Long, Integer> map, TreeNode root, long sum) {
         if (root == null) {
             return 0;
         }
         sum += root.val;
-        long ret;
-
+        int ret;
         ret = map.getOrDefault(sum - targetSum, 0);
         map.put(sum, map.getOrDefault(sum, 0) + 1);
-        ret += (int) dfs(map, root.left, sum);
-        ret += (int) dfs(map, root.right, sum);
+        ret += dfs(map, root.left, sum);
+        ret += dfs(map, root.right, sum);
         map.put(sum, map.getOrDefault(sum, 0) - 1);
         return ret;
     }
